Moved some helpers methods in the wallet controller and wallet manager test files into a static helper methods class
