+++ Stratis.Bitcoin.Tests/Wallet/WalletManagerTest.cs
@@ -19,7 +19,7 @@ using System.Threading.Tasks;
 using Xunit;
 
 namespace Stratis.Bitcoin.Tests.Wallet
-{
+{    
     public class WalletManagerTest : LogsTestBase
     {
         /// <summary>
@@ -290,15 +290,15 @@ namespace Stratis.Bitcoin.Tests.Wallet
                                                   dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
 
             var concurrentChain = new ConcurrentChain(Network.Main);
-            ChainedBlock tip = AppendBlock(null, concurrentChain).ChainedBlock;
+            ChainedBlock tip = WalletTestsHelpers.AppendBlock(null, concurrentChain).ChainedBlock;
 
-            walletManager.Wallets.Add(CreateWallet("wallet1"));
-            walletManager.Wallets.Add(CreateWallet("wallet2"));
+            walletManager.Wallets.Add(WalletTestsHelpers.CreateWallet("wallet1"));
+            walletManager.Wallets.Add(WalletTestsHelpers.CreateWallet("wallet2"));
 
             Parallel.For(0, 500, new ParallelOptions() { MaxDegreeOfParallelism = 10 }, (int iteration) =>
             {
                 walletManager.UpdateLastBlockSyncedHeight(tip);
-                walletManager.Wallets.Add(CreateWallet("wallet"));
+                walletManager.Wallets.Add(WalletTestsHelpers.CreateWallet("wallet"));
                 walletManager.UpdateLastBlockSyncedHeight(tip);
             });
 
@@ -363,9 +363,9 @@ namespace Stratis.Bitcoin.Tests.Wallet
             var passphrase = "this is my magic passphrase";
             var walletName = "mywallet";
 
-            ConcurrentChain chain = PrepareChainWithBlock();
+            ConcurrentChain chain = WalletTestsHelpers.PrepareChainWithBlock();
             // prepare an existing wallet through this manager and delete the file from disk. Return the created wallet object and mnemonic.
-            var deletedWallet = this.CreateWalletOnDiskAndDeleteWallet(dataFolder, password, passphrase, walletName, chain);
+            var deletedWallet = CreateWalletOnDiskAndDeleteWallet(dataFolder, password, passphrase, walletName, chain);
             Assert.False(File.Exists(Path.Combine(dataFolder.WalletPath + $"/{walletName}.wallet.json")));
 
             // create a fresh manager.
@@ -457,7 +457,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             var password = "test";
             var walletName = "mywallet";
 
-            ConcurrentChain chain = PrepareChainWithBlock();
+            ConcurrentChain chain = WalletTestsHelpers.PrepareChainWithBlock();
             // prepare an existing wallet through this manager and delete the file from disk. Return the created wallet object and mnemonic.
             var deletedWallet = this.CreateWalletOnDiskAndDeleteWallet(dataFolder, password, password, walletName, chain);
             Assert.False(File.Exists(Path.Combine(dataFolder.WalletPath + $"/{walletName}.wallet.json")));
@@ -552,10 +552,10 @@ namespace Stratis.Bitcoin.Tests.Wallet
                 dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
 
             // generate 3 wallet with 2 accounts containing 1000 external and 100 internal addresses each.
-            walletManager.Wallets.Add(CreateWallet("wallet1"));
-            walletManager.Wallets.Add(CreateWallet("wallet2"));
-            walletManager.Wallets.Add(CreateWallet("wallet3"));
-            this.AddAddressesToWallet(walletManager, 1000);
+            walletManager.Wallets.Add(WalletTestsHelpers.CreateWallet("wallet1"));
+            walletManager.Wallets.Add(WalletTestsHelpers.CreateWallet("wallet2"));
+            walletManager.Wallets.Add(WalletTestsHelpers.CreateWallet("wallet3"));
+            WalletTestsHelpers.AddAddressesToWallet(walletManager, 1000);
 
             Parallel.For(0, 5000, new ParallelOptions() { MaxDegreeOfParallelism = 10 }, (int iteration) =>
             {
@@ -586,7 +586,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
               dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("testWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("testWallet", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount() { Name = "unused" });
             walletManager.Wallets.Add(wallet);
 
@@ -603,7 +603,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
               dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("testWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("testWallet", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Clear();
             walletManager.Wallets.Add(wallet);
 
@@ -620,7 +620,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
               dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("testWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("testWallet", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount() { Name = "unused" });
             walletManager.Wallets.Add(wallet);
 
@@ -637,7 +637,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
               dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("testWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("testWallet", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Clear();
             walletManager.Wallets.Add(wallet);
 
@@ -654,7 +654,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
               dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("testWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("testWallet", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Clear();
 
             var result = walletManager.CreateNewAccount(wallet, "password");
@@ -677,7 +677,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
               dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("testWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("testWallet", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount() { Name = "unused" });
 
             var result = walletManager.CreateNewAccount(wallet, "password");
@@ -700,7 +700,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             {
                 var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-                var wallet = GenerateBlankWallet("testWallet", "password");
+                var wallet = WalletTestsHelpers.GenerateBlankWallet("testWallet", "password");
                 walletManager.Wallets.Add(wallet);
 
                 var result = walletManager.GetUnusedAddress(new WalletAccountReference("testWallet", "unexistingAccount"));
@@ -722,10 +722,10 @@ namespace Stratis.Bitcoin.Tests.Wallet
         [Fact]
         public void GetUnusedAddressWithWalletHavingUnusedAddressReturnsAddress()
         {
-            var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/CheckWalletBalanceEstimationWithConfirmedTransactions");
+            var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/GetUnusedAddressWithWalletHavingUnusedAddressReturnsAddress");
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                   dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
                 Index = 0,
@@ -761,7 +761,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/GetOrCreateChangeAddressWithWalletHavingUnusedAddressReturnsAddress");
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                 dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
                 Index = 0,
@@ -798,7 +798,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             Directory.CreateDirectory(dataFolder.WalletPath);
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                 dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
 
             var extKey = new ExtKey(Key.Parse(wallet.EncryptedSeed, "password", wallet.Network), wallet.ChainCode);
             var accountExtendedPubKey = extKey.Derive(new KeyPath($"m/44'/0'/0'")).Neuter().ToString(wallet.Network);
@@ -827,7 +827,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             Directory.CreateDirectory(dataFolder.WalletPath);
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                   dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             var extKey = new ExtKey(Key.Parse(wallet.EncryptedSeed, "password", wallet.Network), wallet.ChainCode);
             var accountExtendedPubKey = extKey.Derive(new KeyPath($"m/44'/0'/0'")).Neuter().ToString(wallet.Network);
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
@@ -872,7 +872,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                   new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
                 Index = 0,
@@ -880,12 +880,12 @@ namespace Stratis.Bitcoin.Tests.Wallet
                 HdPath = "m/44'/0'/0'",
                 ExternalAddresses = new List<HdAddress>()
                 {
-                    CreateAddressWithEmptyTransaction(0, "myUsedExternalAddress"),
-                    CreateAddressWithoutTransaction(1, "myUnusedExternalAddress"),
+                    WalletTestsHelpers.CreateAddressWithEmptyTransaction(0, "myUsedExternalAddress"),
+                    WalletTestsHelpers.CreateAddressWithoutTransaction(1, "myUnusedExternalAddress"),
                 },
                 InternalAddresses = new List<HdAddress>() {
-                    CreateAddressWithEmptyTransaction(0, "myUsedInternalAddress"),
-                    CreateAddressWithoutTransaction(1, "myUnusedInternalAddress"),
+                    WalletTestsHelpers.CreateAddressWithEmptyTransaction(0, "myUsedInternalAddress"),
+                    WalletTestsHelpers.CreateAddressWithoutTransaction(1, "myUnusedInternalAddress"),
                 },
                 ExtendedPubKey = "blabla"
             });
@@ -905,7 +905,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                   new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
                 Index = 0,
@@ -913,12 +913,12 @@ namespace Stratis.Bitcoin.Tests.Wallet
                 HdPath = "m/44'/0'/0'",
                 ExternalAddresses = new List<HdAddress>()
                 {
-                    CreateAddressWithEmptyTransaction(0, "myUsedExternalAddress"),
-                    CreateAddressWithoutTransaction(1, "myUnusedExternalAddress"),
+                    WalletTestsHelpers.CreateAddressWithEmptyTransaction(0, "myUsedExternalAddress"),
+                    WalletTestsHelpers.CreateAddressWithoutTransaction(1, "myUnusedExternalAddress"),
                 },
                 InternalAddresses = new List<HdAddress>() {
-                    CreateAddressWithEmptyTransaction(0, "myUsedInternalAddress"),
-                    CreateAddressWithoutTransaction(1, "myUnusedInternalAddress"),
+                    WalletTestsHelpers.CreateAddressWithEmptyTransaction(0, "myUsedInternalAddress"),
+                    WalletTestsHelpers.CreateAddressWithoutTransaction(1, "myUnusedInternalAddress"),
                 },
                 ExtendedPubKey = "blabla"
             });
@@ -938,7 +938,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                   new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
                 Index = 0,
@@ -971,7 +971,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                 new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             walletManager.Wallets.Add(wallet);
 
             var result = walletManager.GetWallet("myWallet");
@@ -995,7 +995,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
              new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount() { Name = "Account 0" });
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount() { Name = "Account 1" });
             wallet.AccountsRoot.Add(new AccountRoot()
@@ -1023,7 +1023,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
              new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet.AccountsRoot.Clear();
             walletManager.Wallets.Add(wallet);
 
@@ -1069,13 +1069,13 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                 new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet.AccountsRoot.ElementAt(0).CoinType = CoinType.Stratis;
             wallet.AccountsRoot.ElementAt(0).LastBlockSyncedHeight = 15;
-            var wallet2 = GenerateBlankWallet("myWallet", "password");
+            var wallet2 = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet2.AccountsRoot.ElementAt(0).CoinType = CoinType.Bitcoin;
             wallet2.AccountsRoot.ElementAt(0).LastBlockSyncedHeight = 20;
-            var wallet3 = GenerateBlankWallet("myWallet", "password");
+            var wallet3 = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet3.AccountsRoot.ElementAt(0).CoinType = CoinType.Bitcoin;
             wallet3.AccountsRoot.ElementAt(0).LastBlockSyncedHeight = 56;
             walletManager.Wallets.Add(wallet);
@@ -1092,7 +1092,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                 new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet.AccountsRoot.ElementAt(0).CoinType = CoinType.Stratis;
             wallet.AccountsRoot.ElementAt(0).LastBlockSyncedHeight = 15;
             wallet.AccountsRoot.Add(new AccountRoot()
@@ -1101,10 +1101,10 @@ namespace Stratis.Bitcoin.Tests.Wallet
                 LastBlockSyncedHeight = 12
             });
 
-            var wallet2 = GenerateBlankWallet("myWallet", "password");
+            var wallet2 = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet2.AccountsRoot.ElementAt(0).CoinType = CoinType.Bitcoin;
             wallet2.AccountsRoot.ElementAt(0).LastBlockSyncedHeight = 20;
-            var wallet3 = GenerateBlankWallet("myWallet", "password");
+            var wallet3 = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet3.AccountsRoot.ElementAt(0).CoinType = CoinType.Bitcoin;
             wallet3.AccountsRoot.ElementAt(0).LastBlockSyncedHeight = 56;
             walletManager.Wallets.Add(wallet);
@@ -1121,7 +1121,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                 new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet.AccountsRoot.ElementAt(0).CoinType = CoinType.Stratis;
             walletManager.Wallets.Add(wallet);
 
@@ -1155,15 +1155,15 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                 new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet.AccountsRoot.ElementAt(0).CoinType = CoinType.Stratis;
             wallet.AccountsRoot.ElementAt(0).LastBlockSyncedHeight = 15;
             wallet.AccountsRoot.ElementAt(0).LastBlockSyncedHash = new uint256(15);
-            var wallet2 = GenerateBlankWallet("myWallet", "password");
+            var wallet2 = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet2.AccountsRoot.ElementAt(0).CoinType = CoinType.Bitcoin;
             wallet2.AccountsRoot.ElementAt(0).LastBlockSyncedHeight = 20;
             wallet2.AccountsRoot.ElementAt(0).LastBlockSyncedHash = new uint256(20);
-            var wallet3 = GenerateBlankWallet("myWallet", "password");
+            var wallet3 = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet3.AccountsRoot.ElementAt(0).CoinType = CoinType.Bitcoin;
             wallet3.AccountsRoot.ElementAt(0).LastBlockSyncedHeight = 56;
             wallet3.AccountsRoot.ElementAt(0).LastBlockSyncedHash = new uint256(56);
@@ -1181,7 +1181,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                 new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet.AccountsRoot.ElementAt(0).CoinType = CoinType.Stratis;
             wallet.AccountsRoot.ElementAt(0).LastBlockSyncedHeight = 15;
             wallet.AccountsRoot.ElementAt(0).LastBlockSyncedHash = new uint256(15);
@@ -1192,11 +1192,11 @@ namespace Stratis.Bitcoin.Tests.Wallet
                 LastBlockSyncedHash = new uint256(12)
             });
 
-            var wallet2 = GenerateBlankWallet("myWallet", "password");
+            var wallet2 = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet2.AccountsRoot.ElementAt(0).CoinType = CoinType.Bitcoin;
             wallet2.AccountsRoot.ElementAt(0).LastBlockSyncedHeight = 20;
             wallet2.AccountsRoot.ElementAt(0).LastBlockSyncedHash = new uint256(20);
-            var wallet3 = GenerateBlankWallet("myWallet", "password");
+            var wallet3 = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet3.AccountsRoot.ElementAt(0).CoinType = CoinType.Bitcoin;
             wallet3.AccountsRoot.ElementAt(0).LastBlockSyncedHeight = 56;
             wallet3.AccountsRoot.ElementAt(0).LastBlockSyncedHash = new uint256(56);
@@ -1216,7 +1216,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             {
                 var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                     new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-                var wallet = GenerateBlankWallet("myWallet", "password");
+                var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
                 wallet.AccountsRoot.ElementAt(0).CoinType = CoinType.Stratis;
                 walletManager.Wallets.Add(wallet);
 
@@ -1227,14 +1227,14 @@ namespace Stratis.Bitcoin.Tests.Wallet
         [Fact]
         public void GetSpendableTransactionsWithChainOfHeightZeroReturnsNoTransactions()
         {
-            var chain = GenerateChainWithHeight(0, Network.Main);
+            var chain = WalletTestsHelpers.GenerateChainWithHeight(0, Network.Main);
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, chain, NodeSettings.Default(),
                     new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
-                ExternalAddresses = CreateUnspentTransactionsOfBlockHeights(Network.Main, 1, 9, 10),
-                InternalAddresses = CreateUnspentTransactionsOfBlockHeights(Network.Main, 2, 9, 10)
+                ExternalAddresses = WalletTestsHelpers.CreateUnspentTransactionsOfBlockHeights(Network.Main, 1, 9, 10),
+                InternalAddresses = WalletTestsHelpers.CreateUnspentTransactionsOfBlockHeights(Network.Main, 2, 9, 10)
             });
 
             walletManager.Wallets.Add(wallet);
@@ -1251,15 +1251,15 @@ namespace Stratis.Bitcoin.Tests.Wallet
         [Fact]
         public void GetSpendableTransactionsReturnsTransactionsGivenBlockHeight()
         {
-            var chain = GenerateChainWithHeight(10, Network.Main);
+            var chain = WalletTestsHelpers.GenerateChainWithHeight(10, Network.Main);
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, chain, NodeSettings.Default(),
                     new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet1", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
                 Name = "First expectation",
-                ExternalAddresses = CreateUnspentTransactionsOfBlockHeights(Network.Main, 1, 9, 10),
-                InternalAddresses = CreateUnspentTransactionsOfBlockHeights(Network.Main, 2, 9, 10)
+                ExternalAddresses = WalletTestsHelpers.CreateUnspentTransactionsOfBlockHeights(Network.Main, 1, 9, 10),
+                InternalAddresses = WalletTestsHelpers.CreateUnspentTransactionsOfBlockHeights(Network.Main, 2, 9, 10)
             });
 
             wallet.AccountsRoot.Add(new AccountRoot()
@@ -1268,26 +1268,26 @@ namespace Stratis.Bitcoin.Tests.Wallet
                 Accounts = new List<HdAccount>()
                 {
                     new HdAccount() {
-                        ExternalAddresses = CreateUnspentTransactionsOfBlockHeights(Network.StratisMain, 8,9,10),
-                        InternalAddresses = CreateUnspentTransactionsOfBlockHeights(Network.StratisMain, 8,9,10)
+                        ExternalAddresses = WalletTestsHelpers.CreateUnspentTransactionsOfBlockHeights(Network.StratisMain, 8,9,10),
+                        InternalAddresses = WalletTestsHelpers.CreateUnspentTransactionsOfBlockHeights(Network.StratisMain, 8,9,10)
                     }
                 }
             });
 
-            var wallet2 = GenerateBlankWallet("myWallet2", "password");
+            var wallet2 = WalletTestsHelpers.GenerateBlankWallet("myWallet2", "password");
             wallet2.AccountsRoot.ElementAt(0).CoinType = CoinType.Stratis;
             wallet2.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
-                ExternalAddresses = CreateUnspentTransactionsOfBlockHeights(Network.StratisMain, 1, 3, 5, 7, 9, 10),
-                InternalAddresses = CreateUnspentTransactionsOfBlockHeights(Network.StratisMain, 2, 4, 6, 8, 9, 10)
+                ExternalAddresses = WalletTestsHelpers.CreateUnspentTransactionsOfBlockHeights(Network.StratisMain, 1, 3, 5, 7, 9, 10),
+                InternalAddresses = WalletTestsHelpers.CreateUnspentTransactionsOfBlockHeights(Network.StratisMain, 2, 4, 6, 8, 9, 10)
             });
 
-            var wallet3 = GenerateBlankWallet("myWallet3", "password");
+            var wallet3 = WalletTestsHelpers.GenerateBlankWallet("myWallet3", "password");
             wallet3.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
                 Name = "Second expectation",
-                ExternalAddresses = CreateUnspentTransactionsOfBlockHeights(Network.Main, 5, 9, 10),
-                InternalAddresses = CreateUnspentTransactionsOfBlockHeights(Network.Main, 6, 9, 10)
+                ExternalAddresses = WalletTestsHelpers.CreateUnspentTransactionsOfBlockHeights(Network.Main, 5, 9, 10),
+                InternalAddresses = WalletTestsHelpers.CreateUnspentTransactionsOfBlockHeights(Network.Main, 6, 9, 10)
             });
 
             walletManager.Wallets.Add(wallet);
@@ -1320,15 +1320,15 @@ namespace Stratis.Bitcoin.Tests.Wallet
         [Fact]
         public void GetSpendableTransactionsWithSpentTransactionsReturnsSpendableTransactionsGivenBlockHeight()
         {
-            var chain = GenerateChainWithHeight(10, Network.Main);
+            var chain = WalletTestsHelpers.GenerateChainWithHeight(10, Network.Main);
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, chain, NodeSettings.Default(),
                     new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet1", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
                 Name = "First expectation",
-                ExternalAddresses = CreateUnspentTransactionsOfBlockHeights(Network.Main, 1, 9, 10).Concat(CreateSpentTransactionsOfBlockHeights(Network.Main, 1, 9, 10)).ToList(),
-                InternalAddresses = CreateUnspentTransactionsOfBlockHeights(Network.Main, 2, 9, 10).Concat(CreateSpentTransactionsOfBlockHeights(Network.Main, 2, 9, 10)).ToList()
+                ExternalAddresses = WalletTestsHelpers.CreateUnspentTransactionsOfBlockHeights(Network.Main, 1, 9, 10).Concat(WalletTestsHelpers.CreateSpentTransactionsOfBlockHeights(Network.Main, 1, 9, 10)).ToList(),
+                InternalAddresses = WalletTestsHelpers.CreateUnspentTransactionsOfBlockHeights(Network.Main, 2, 9, 10).Concat(WalletTestsHelpers.CreateSpentTransactionsOfBlockHeights(Network.Main, 2, 9, 10)).ToList()
             });
 
             walletManager.Wallets.Add(wallet);
@@ -1365,7 +1365,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             Assert.Throws<WalletException>(() =>
             {
-                var chain = GenerateChainWithHeight(10, Network.Main);
+                var chain = WalletTestsHelpers.GenerateChainWithHeight(10, Network.Main);
                 var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, chain, NodeSettings.Default(),
                         new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
 
@@ -1376,16 +1376,16 @@ namespace Stratis.Bitcoin.Tests.Wallet
         [Fact]
         public void GetSpendableTransactionsWithoutWalletsOfWalletManagerCoinTypeReturnsEmptyList()
         {
-            var chain = GenerateChainWithHeight(10, Network.Main);
+            var chain = WalletTestsHelpers.GenerateChainWithHeight(10, Network.Main);
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, chain, NodeSettings.Default(),
                     new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
 
-            var wallet = GenerateBlankWallet("myWallet2", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet2", "password");
             wallet.AccountsRoot.ElementAt(0).CoinType = CoinType.Stratis;
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
-                ExternalAddresses = CreateUnspentTransactionsOfBlockHeights(Network.StratisMain, 1, 3, 5, 7, 9, 10),
-                InternalAddresses = CreateUnspentTransactionsOfBlockHeights(Network.StratisMain, 2, 4, 6, 8, 9, 10)
+                ExternalAddresses = WalletTestsHelpers.CreateUnspentTransactionsOfBlockHeights(Network.StratisMain, 1, 3, 5, 7, 9, 10),
+                InternalAddresses = WalletTestsHelpers.CreateUnspentTransactionsOfBlockHeights(Network.StratisMain, 2, 4, 6, 8, 9, 10)
             });
             walletManager.Wallets.Add(wallet);
 
@@ -1397,15 +1397,15 @@ namespace Stratis.Bitcoin.Tests.Wallet
         [Fact]
         public void GetSpendableTransactionsWithOnlySpentTransactionsReturnsEmptyList()
         {
-            var chain = GenerateChainWithHeight(10, Network.Main);
+            var chain = WalletTestsHelpers.GenerateChainWithHeight(10, Network.Main);
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, chain, NodeSettings.Default(),
                     new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var wallet = GenerateBlankWallet("myWallet1", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
                 Name = "First expectation",
-                ExternalAddresses = CreateSpentTransactionsOfBlockHeights(Network.Main, 1, 9, 10),
-                InternalAddresses = CreateSpentTransactionsOfBlockHeights(Network.Main, 2, 9, 10)
+                ExternalAddresses = WalletTestsHelpers.CreateSpentTransactionsOfBlockHeights(Network.Main, 1, 9, 10),
+                InternalAddresses = WalletTestsHelpers.CreateSpentTransactionsOfBlockHeights(Network.Main, 2, 9, 10)
             });
 
             walletManager.Wallets.Add(wallet);
@@ -1433,7 +1433,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                       new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-            var data = GenerateBlankWalletWithExtKey("myWallet", "password");
+            var data = WalletTestsHelpers.GenerateBlankWalletWithExtKey("myWallet", "password");
 
             var address = new HdAddress
             {
@@ -1464,7 +1464,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             {
                 var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                           new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
-                var data = GenerateBlankWalletWithExtKey("myWallet", "password");
+                var data = WalletTestsHelpers.GenerateBlankWalletWithExtKey("myWallet", "password");
 
                 var address = new HdAddress
                 {
@@ -1491,11 +1491,11 @@ namespace Stratis.Bitcoin.Tests.Wallet
             var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/LoadWalletWithExistingWalletLoadsWalletOntoManager");
             Directory.CreateDirectory(dataFolder.WalletPath);
 
-            var wallet = GenerateBlankWallet("myWallet1", "password");
-            var accountKeys = GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
-            var spendingKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
-            var destinationKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
-            var changeKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
+            var accountKeys = WalletTestsHelpers.GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
+            var spendingKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
+            var destinationKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
+            var changeKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
 
             var spendingAddress = new HdAddress()
             {
@@ -1531,8 +1531,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
             };
 
             //Generate a spendable transaction
-            var chainInfo = CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
-            TransactionData spendingTransaction = CreateTransactionDataFromFirstBlock(chainInfo);
+            var chainInfo = WalletTestsHelpers.CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
+            TransactionData spendingTransaction = WalletTestsHelpers.CreateTransactionDataFromFirstBlock(chainInfo);
             spendingAddress.Transactions.Add(spendingTransaction);
 
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
@@ -1546,7 +1546,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             });
 
             // setup a payment to yourself
-            var transaction = SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
+            var transaction = WalletTestsHelpers.SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
 
             var walletFeePolicy = new Mock<IWalletFeePolicy>();
             walletFeePolicy.Setup(w => w.GetMinimumFee(258, 50))
@@ -1583,11 +1583,11 @@ namespace Stratis.Bitcoin.Tests.Wallet
             var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/LoadWalletWithExistingWalletLoadsWalletOntoManager");
             Directory.CreateDirectory(dataFolder.WalletPath);
 
-            var wallet = GenerateBlankWallet("myWallet1", "password");
-            var accountKeys = GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
-            var spendingKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
-            var destinationKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
-            var changeKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
+            var accountKeys = WalletTestsHelpers.GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
+            var spendingKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
+            var destinationKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
+            var changeKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
 
             var spendingAddress = new HdAddress()
             {
@@ -1623,8 +1623,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
             };
 
             //Generate a spendable transaction
-            var chainInfo = CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
-            TransactionData spendingTransaction = CreateTransactionDataFromFirstBlock(chainInfo);
+            var chainInfo = WalletTestsHelpers.CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
+            TransactionData spendingTransaction = WalletTestsHelpers.CreateTransactionDataFromFirstBlock(chainInfo);
             spendingAddress.Transactions.Add(spendingTransaction);
 
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
@@ -1638,7 +1638,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             });
 
             // setup a payment to yourself
-            var transaction = SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
+            var transaction = WalletTestsHelpers.SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
             transaction.Outputs.ElementAt(1).Value = Money.Zero;
             transaction.Outputs.ElementAt(1).ScriptPubKey = Script.Empty;
 
@@ -1672,11 +1672,11 @@ namespace Stratis.Bitcoin.Tests.Wallet
             var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/LoadWalletWithExistingWalletLoadsWalletOntoManager");
             Directory.CreateDirectory(dataFolder.WalletPath);
 
-            var wallet = GenerateBlankWallet("myWallet1", "password");
-            var accountKeys = GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
-            var spendingKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
-            var changeKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
-            var destinationKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/1");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
+            var accountKeys = WalletTestsHelpers.GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
+            var spendingKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
+            var changeKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
+            var destinationKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/1");
 
             var spendingAddress = new HdAddress()
             {
@@ -1712,8 +1712,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
             };
 
             //Generate a spendable transaction
-            var chainInfo = CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
-            TransactionData spendingTransaction = CreateTransactionDataFromFirstBlock(chainInfo);
+            var chainInfo = WalletTestsHelpers.CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
+            TransactionData spendingTransaction = WalletTestsHelpers.CreateTransactionDataFromFirstBlock(chainInfo);
             spendingAddress.Transactions.Add(spendingTransaction);
 
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
@@ -1727,7 +1727,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             });
 
             // setup a payment to yourself
-            var transaction = SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
+            var transaction = WalletTestsHelpers.SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
 
             var walletFeePolicy = new Mock<IWalletFeePolicy>();
             walletFeePolicy.Setup(w => w.GetMinimumFee(258, 50))
@@ -1766,11 +1766,11 @@ namespace Stratis.Bitcoin.Tests.Wallet
             var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/LoadWalletWithExistingWalletLoadsWalletOntoManager");
             Directory.CreateDirectory(dataFolder.WalletPath);
 
-            var wallet = GenerateBlankWallet("myWallet1", "password");
-            var accountKeys = GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
-            var spendingKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
-            var destinationKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
-            var changeKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
+            var accountKeys = WalletTestsHelpers.GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
+            var spendingKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
+            var destinationKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
+            var changeKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
 
             var spendingAddress = new HdAddress()
             {
@@ -1806,8 +1806,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
             };
 
             //Generate a spendable transaction
-            var chainInfo = CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
-            TransactionData spendingTransaction = CreateTransactionDataFromFirstBlock(chainInfo);
+            var chainInfo = WalletTestsHelpers.CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
+            TransactionData spendingTransaction = WalletTestsHelpers.CreateTransactionDataFromFirstBlock(chainInfo);
             spendingAddress.Transactions.Add(spendingTransaction);
 
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
@@ -1821,7 +1821,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             });
 
             // setup a payment to yourself
-            var transaction = SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
+            var transaction = WalletTestsHelpers.SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
 
             var walletFeePolicy = new Mock<IWalletFeePolicy>();
             walletFeePolicy.Setup(w => w.GetMinimumFee(258, 50))
@@ -1831,7 +1831,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
                   dataFolder, walletFeePolicy.Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
             walletManager.Wallets.Add(wallet);
 
-            var block = AppendTransactionInNewBlockToChain(chainInfo.chain, transaction);
+            var block = WalletTestsHelpers.AppendTransactionInNewBlockToChain(chainInfo.chain, transaction);
 
             var blockHeight = chainInfo.chain.GetBlock(block.GetHash()).Height;
             walletManager.ProcessTransaction(transaction, blockHeight);
@@ -1864,11 +1864,11 @@ namespace Stratis.Bitcoin.Tests.Wallet
             var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/LoadWalletWithExistingWalletLoadsWalletOntoManager");
             Directory.CreateDirectory(dataFolder.WalletPath);
 
-            var wallet = GenerateBlankWallet("myWallet1", "password");
-            var accountKeys = GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
-            var spendingKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
-            var destinationKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
-            var changeKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
+            var accountKeys = WalletTestsHelpers.GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
+            var spendingKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
+            var destinationKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
+            var changeKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
 
             var spendingAddress = new HdAddress()
             {
@@ -1904,8 +1904,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
             };
 
             //Generate a spendable transaction
-            var chainInfo = CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
-            TransactionData spendingTransaction = CreateTransactionDataFromFirstBlock(chainInfo);
+            var chainInfo = WalletTestsHelpers.CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
+            TransactionData spendingTransaction = WalletTestsHelpers.CreateTransactionDataFromFirstBlock(chainInfo);
             spendingAddress.Transactions.Add(spendingTransaction);
 
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
@@ -1919,7 +1919,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             });
 
             // setup a payment to yourself
-            var transaction = SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
+            var transaction = WalletTestsHelpers.SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
 
             var walletFeePolicy = new Mock<IWalletFeePolicy>();
             walletFeePolicy.Setup(w => w.GetMinimumFee(258, 50))
@@ -1929,7 +1929,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
                   dataFolder, walletFeePolicy.Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
             walletManager.Wallets.Add(wallet);
 
-            var block = AppendTransactionInNewBlockToChain(chainInfo.chain, transaction);
+            var block = WalletTestsHelpers.AppendTransactionInNewBlockToChain(chainInfo.chain, transaction);
 
             walletManager.ProcessTransaction(transaction, block: block);
 
@@ -1961,11 +1961,11 @@ namespace Stratis.Bitcoin.Tests.Wallet
             var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/LoadWalletWithExistingWalletLoadsWalletOntoManager");
             Directory.CreateDirectory(dataFolder.WalletPath);
 
-            var wallet = GenerateBlankWallet("myWallet1", "password");
-            var accountKeys = GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
-            var spendingKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
-            var destinationKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
-            var changeKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
+            var accountKeys = WalletTestsHelpers.GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
+            var spendingKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
+            var destinationKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
+            var changeKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
 
             var spendingAddress = new HdAddress()
             {
@@ -2001,8 +2001,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
             };
 
             //Generate a spendable transaction
-            var chainInfo = CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
-            TransactionData spendingTransaction = CreateTransactionDataFromFirstBlock(chainInfo);
+            var chainInfo = WalletTestsHelpers.CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
+            TransactionData spendingTransaction = WalletTestsHelpers.CreateTransactionDataFromFirstBlock(chainInfo);
             spendingAddress.Transactions.Add(spendingTransaction);
 
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
@@ -2016,7 +2016,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             });
 
             // setup a payment to yourself
-            var transaction = SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
+            var transaction = WalletTestsHelpers.SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
             transaction.Outputs.ElementAt(1).Value = Money.Zero;
             transaction.Outputs.ElementAt(1).ScriptPubKey = Script.Empty;
 
@@ -2074,11 +2074,11 @@ namespace Stratis.Bitcoin.Tests.Wallet
             var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/LoadWalletWithExistingWalletLoadsWalletOntoManager");
             Directory.CreateDirectory(dataFolder.WalletPath);
 
-            var wallet = GenerateBlankWallet("myWallet1", "password");
-            var accountKeys = GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
-            var spendingKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
-            var destinationKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
-            var changeKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
+            var accountKeys = WalletTestsHelpers.GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
+            var spendingKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
+            var destinationKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
+            var changeKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
 
             var spendingAddress = new HdAddress()
             {
@@ -2114,8 +2114,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
             };
 
             //Generate a spendable transaction
-            var chainInfo = CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
-            TransactionData spendingTransaction = CreateTransactionDataFromFirstBlock(chainInfo);
+            var chainInfo = WalletTestsHelpers.CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
+            TransactionData spendingTransaction = WalletTestsHelpers.CreateTransactionDataFromFirstBlock(chainInfo);
             spendingAddress.Transactions.Add(spendingTransaction);
 
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
@@ -2129,7 +2129,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             });
 
             // setup a payment to yourself
-            var transaction = SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
+            var transaction = WalletTestsHelpers.SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
             transaction.Outputs.ElementAt(1).Value = Money.Zero;
             transaction.Outputs.ElementAt(1).ScriptPubKey = Script.Empty;
 
@@ -2195,11 +2195,11 @@ namespace Stratis.Bitcoin.Tests.Wallet
             var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/LoadWalletWithExistingWalletLoadsWalletOntoManager");
             Directory.CreateDirectory(dataFolder.WalletPath);
 
-            var wallet = GenerateBlankWallet("myWallet1", "password");
-            var accountKeys = GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
-            var spendingKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
-            var destinationKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
-            var changeKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
+            var accountKeys = WalletTestsHelpers.GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
+            var spendingKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
+            var destinationKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
+            var changeKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
 
             var spendingAddress = new HdAddress()
             {
@@ -2235,8 +2235,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
             };
 
             //Generate a spendable transaction
-            var chainInfo = CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
-            TransactionData spendingTransaction = CreateTransactionDataFromFirstBlock(chainInfo);
+            var chainInfo = WalletTestsHelpers.CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
+            TransactionData spendingTransaction = WalletTestsHelpers.CreateTransactionDataFromFirstBlock(chainInfo);
             spendingAddress.Transactions.Add(spendingTransaction);
 
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
@@ -2250,7 +2250,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
             });
 
             // setup a payment to yourself
-            var transaction = SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
+            var transaction = WalletTestsHelpers.SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
             transaction.Outputs.ElementAt(1).Value = Money.Zero;
             transaction.Outputs.ElementAt(1).ScriptPubKey = Script.Empty;
 
@@ -2302,14 +2302,14 @@ namespace Stratis.Bitcoin.Tests.Wallet
         public void RemoveBlocksRemovesTransactionsWithHigherBlockHeightAndUpdatesLastSyncedBlockHeight()
         {
             var concurrentchain = new ConcurrentChain(Network.Main);
-            var chainedBlock = AppendBlock(null, concurrentchain).ChainedBlock;
+            var chainedBlock = WalletTestsHelpers.AppendBlock(null, concurrentchain).ChainedBlock;
 
-            var wallet = GenerateBlankWallet("myWallet1", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
                 Name = "First account",
-                ExternalAddresses = CreateSpentTransactionsOfBlockHeights(Network.Main, 1, 2, 3).ToList(),
-                InternalAddresses = CreateSpentTransactionsOfBlockHeights(Network.Main, 1, 2, 3).ToList()
+                ExternalAddresses = WalletTestsHelpers.CreateSpentTransactionsOfBlockHeights(Network.Main, 1, 2, 3).ToList(),
+                InternalAddresses = WalletTestsHelpers.CreateSpentTransactionsOfBlockHeights(Network.Main, 1, 2, 3).ToList()
             });
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
@@ -2333,7 +2333,7 @@ namespace Stratis.Bitcoin.Tests.Wallet
         public void ProcessBlockWithoutWalletsSetsWalletTipToBlockHash()
         {
             var concurrentchain = new ConcurrentChain(Network.Main);
-            var blockResult = AppendBlock(null, concurrentchain);
+            var blockResult = WalletTestsHelpers.AppendBlock(null, concurrentchain);
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                 new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
@@ -2349,11 +2349,11 @@ namespace Stratis.Bitcoin.Tests.Wallet
             var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/LoadWalletWithExistingWalletLoadsWalletOntoManager");
             Directory.CreateDirectory(dataFolder.WalletPath);
 
-            var wallet = GenerateBlankWallet("myWallet1", "password");
-            var accountKeys = GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
-            var spendingKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
-            var destinationKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
-            var changeKeys = GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
+            var accountKeys = WalletTestsHelpers.GenerateAccountKeys(wallet, "password", "m/44'/0'/0'");
+            var spendingKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/0");
+            var destinationKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "0/1");
+            var changeKeys = WalletTestsHelpers.GenerateAddressKeys(wallet, accountKeys.ExtPubKey, "1/0");
 
             var spendingAddress = new HdAddress()
             {
@@ -2389,14 +2389,14 @@ namespace Stratis.Bitcoin.Tests.Wallet
             };
 
             //Generate a spendable transaction
-            var chainInfo = CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
+            var chainInfo = WalletTestsHelpers.CreateChainAndCreateFirstBlockWithPaymentToAddress(wallet.Network, spendingAddress);
 
-            TransactionData spendingTransaction = CreateTransactionDataFromFirstBlock(chainInfo);
+            TransactionData spendingTransaction = WalletTestsHelpers.CreateTransactionDataFromFirstBlock(chainInfo);
             spendingAddress.Transactions.Add(spendingTransaction);
 
             // setup a payment to yourself in a new block.
-            var transaction = SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
-            var block = AppendTransactionInNewBlockToChain(chainInfo.chain, transaction);
+            var transaction = WalletTestsHelpers.SetupValidTransaction(wallet, "password", spendingAddress, destinationKeys.PubKey, changeAddress, new Money(7500), new Money(5000));
+            var block = WalletTestsHelpers.AppendTransactionInNewBlockToChain(chainInfo.chain, transaction);
 
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
@@ -2453,10 +2453,10 @@ namespace Stratis.Bitcoin.Tests.Wallet
                 var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/LoadWalletWithExistingWalletLoadsWalletOntoManager");
                 Directory.CreateDirectory(dataFolder.WalletPath);
 
-                var wallet = GenerateBlankWallet("myWallet1", "password");
+                var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
 
                 ConcurrentChain chain = new ConcurrentChain(wallet.Network.GetGenesis().Header);
-                var chainResult = AppendBlock(chain.Genesis, chain);
+                var chainResult = WalletTestsHelpers.AppendBlock(chain.Genesis, chain);
 
                 var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, chain, NodeSettings.Default(),
                       dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
@@ -2476,11 +2476,11 @@ namespace Stratis.Bitcoin.Tests.Wallet
                 var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/LoadWalletWithExistingWalletLoadsWalletOntoManager");
                 Directory.CreateDirectory(dataFolder.WalletPath);
 
-                var wallet = GenerateBlankWallet("myWallet1", "password");
+                var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
 
                 ConcurrentChain chain = new ConcurrentChain(wallet.Network.GetGenesis().Header);
-                var chainResult = AppendBlock(chain.Genesis, chain);
-                var chainResult2 = AppendBlock(chainResult.ChainedBlock, chain);
+                var chainResult = WalletTestsHelpers.AppendBlock(chain.Genesis, chain);
+                var chainResult2 = WalletTestsHelpers.AppendBlock(chainResult.ChainedBlock, chain);
 
                 var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, chain, NodeSettings.Default(),
                       dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
@@ -2502,8 +2502,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
                 dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
 
             // generate 3 wallet with 2 accounts containing 1000 external and 100 internal addresses each.
-            walletManager.Wallets.Add(CreateWallet("wallet1"));
-            this.AddAddressesToWallet(walletManager, 1000);
+            walletManager.Wallets.Add(WalletTestsHelpers.CreateWallet("wallet1"));
+            WalletTestsHelpers.AddAddressesToWallet(walletManager, 1000);
 
             var firstAccount = walletManager.Wallets.First().AccountsRoot.First().Accounts.First();
 
@@ -2528,8 +2528,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
                 dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
 
             // generate 3 wallet with 2 accounts containing 1000 external and 100 internal addresses each.
-            walletManager.Wallets.Add(CreateWallet("wallet1"));
-            this.AddAddressesToWallet(walletManager, 1000);
+            walletManager.Wallets.Add(WalletTestsHelpers.CreateWallet("wallet1"));
+            WalletTestsHelpers.AddAddressesToWallet(walletManager, 1000);
 
             var firstAccount = walletManager.Wallets.First().AccountsRoot.First().Accounts.First();
 
@@ -2554,8 +2554,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
                 dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
 
             // generate 3 wallet with 2 accounts containing 1000 external and 100 internal addresses each.
-            walletManager.Wallets.Add(CreateWallet("wallet1"));
-            this.AddAddressesToWallet(walletManager, 1000);
+            walletManager.Wallets.Add(WalletTestsHelpers.CreateWallet("wallet1"));
+            WalletTestsHelpers.AddAddressesToWallet(walletManager, 1000);
 
             var firstAccount = walletManager.Wallets.First().AccountsRoot.First().Accounts.First();
 
@@ -2580,8 +2580,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
                 dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
 
             // generate 3 wallet with 2 accounts containing 1000 external and 100 internal addresses each.
-            walletManager.Wallets.Add(CreateWallet("wallet1"));
-            this.AddAddressesToWallet(walletManager, 1000);
+            walletManager.Wallets.Add(WalletTestsHelpers.CreateWallet("wallet1"));
+            WalletTestsHelpers.AddAddressesToWallet(walletManager, 1000);
 
             var firstAccount = walletManager.Wallets.First().AccountsRoot.First().Accounts.First();
 
@@ -2612,8 +2612,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
                 dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
 
             // generate 3 wallet with 2 accounts containing 1000 external and 100 internal addresses each.
-            walletManager.Wallets.Add(CreateWallet("wallet1"));
-            this.AddAddressesToWallet(walletManager, 1000);
+            walletManager.Wallets.Add(WalletTestsHelpers.CreateWallet("wallet1"));
+            WalletTestsHelpers.AddAddressesToWallet(walletManager, 1000);
 
             var firstAccount = walletManager.Wallets.First().AccountsRoot.First().Accounts.First();
 
@@ -2639,8 +2639,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/SaveToFileWithoutWalletParameterSavesAllWalletsOnManagerToDisk");
             Directory.CreateDirectory(dataFolder.WalletPath);
-            var wallet = GenerateBlankWallet("wallet1", "test");
-            var wallet2 = GenerateBlankWallet("wallet2", "test");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("wallet1", "test");
+            var wallet2 = WalletTestsHelpers.GenerateBlankWallet("wallet2", "test");
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                 dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
@@ -2675,8 +2675,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/SaveToFileWithWalletParameterSavesGivenWalletToDisk");
             Directory.CreateDirectory(dataFolder.WalletPath);
-            var wallet = GenerateBlankWallet("wallet1", "test");
-            var wallet2 = GenerateBlankWallet("wallet2", "test");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("wallet1", "test");
+            var wallet2 = WalletTestsHelpers.GenerateBlankWallet("wallet2", "test");
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                 dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
@@ -2713,8 +2713,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
         [Fact]
         public void GetWalletsReturnsLoadedWalletNames()
         {
-            var wallet = GenerateBlankWallet("wallet1", "test");
-            var wallet2 = GenerateBlankWallet("wallet2", "test");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("wallet1", "test");
+            var wallet2 = WalletTestsHelpers.GenerateBlankWallet("wallet2", "test");
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                 new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
@@ -2742,12 +2742,12 @@ namespace Stratis.Bitcoin.Tests.Wallet
         [Fact]
         public void LoadKeysLookupWithKeysLoadsKeyLookup()
         {
-            var wallet = GenerateBlankWallet("myWallet1", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
             wallet.AccountsRoot.ElementAt(0).Accounts.Add(new HdAccount()
             {
                 Name = "First account",
-                ExternalAddresses = CreateSpentTransactionsOfBlockHeights(Network.Main, 1, 2, 3).ToList(),
-                InternalAddresses = CreateSpentTransactionsOfBlockHeights(Network.Main, 1, 2, 3).ToList()
+                ExternalAddresses = WalletTestsHelpers.CreateSpentTransactionsOfBlockHeights(Network.Main, 1, 2, 3).ToList(),
+                InternalAddresses = WalletTestsHelpers.CreateSpentTransactionsOfBlockHeights(Network.Main, 1, 2, 3).ToList()
             });
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
@@ -2803,8 +2803,8 @@ namespace Stratis.Bitcoin.Tests.Wallet
         {
             var dataFolder = AssureEmptyDirAsDataFolder("TestData/WalletManagerTest/DisposeSavesWallets");
             Directory.CreateDirectory(dataFolder.WalletPath);
-            var wallet = GenerateBlankWallet("wallet1", "test");
-            var wallet2 = GenerateBlankWallet("wallet2", "test");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("wallet1", "test");
+            var wallet2 = WalletTestsHelpers.GenerateBlankWallet("wallet2", "test");
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, new Mock<ConcurrentChain>().Object, NodeSettings.Default(),
                 dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
@@ -2837,11 +2837,11 @@ namespace Stratis.Bitcoin.Tests.Wallet
         [Fact]
         public void UpdateLastBlockSyncedHeightWithChainedBlockUpdatesWallets()
         {
-            var wallet = GenerateBlankWallet("myWallet1", "password");
-            var wallet2 = GenerateBlankWallet("myWallet2", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
+            var wallet2 = WalletTestsHelpers.GenerateBlankWallet("myWallet2", "password");
 
             ConcurrentChain chain = new ConcurrentChain(wallet.Network.GetGenesis().Header);
-            var chainedBlock = AppendBlock(chain.Genesis, chain).ChainedBlock;
+            var chainedBlock = WalletTestsHelpers.AppendBlock(chain.Genesis, chain).ChainedBlock;
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, chain, NodeSettings.Default(),
                   new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
@@ -2863,11 +2863,11 @@ namespace Stratis.Bitcoin.Tests.Wallet
         [Fact]
         public void UpdateLastBlockSyncedHeightWithWalletAndChainedBlockUpdatesGivenWallet()
         {
-            var wallet = GenerateBlankWallet("myWallet1", "password");
-            var wallet2 = GenerateBlankWallet("myWallet2", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
+            var wallet2 = WalletTestsHelpers.GenerateBlankWallet("myWallet2", "password");
 
             ConcurrentChain chain = new ConcurrentChain(wallet.Network.GetGenesis().Header);
-            var chainedBlock = AppendBlock(chain.Genesis, chain).ChainedBlock;
+            var chainedBlock = WalletTestsHelpers.AppendBlock(chain.Genesis, chain).ChainedBlock;
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, chain, NodeSettings.Default(),
                   new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
@@ -2890,11 +2890,11 @@ namespace Stratis.Bitcoin.Tests.Wallet
         [Fact]
         public void UpdateLastBlockSyncedHeightWithWalletAccountRootOfDifferentCoinTypeDoesNotUpdateLastSyncedInformation()
         {
-            var wallet = GenerateBlankWallet("myWallet1", "password");
+            var wallet = WalletTestsHelpers.GenerateBlankWallet("myWallet1", "password");
             wallet.AccountsRoot.ElementAt(0).CoinType = CoinType.Stratis;            
 
             ConcurrentChain chain = new ConcurrentChain(wallet.Network.GetGenesis().Header);
-            var chainedBlock = AppendBlock(chain.Genesis, chain).ChainedBlock;
+            var chainedBlock = WalletTestsHelpers.AppendBlock(chain.Genesis, chain).ChainedBlock;
 
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.Main, chain, NodeSettings.Default(),
                   new DataFolder(new NodeSettings() { DataDir = "/TestData/WalletManagerTest" }), new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
@@ -2909,258 +2909,10 @@ namespace Stratis.Bitcoin.Tests.Wallet
             Assert.NotEqual(chainedBlock.HashBlock, walletManager.WalletTipHash);
         }
 
-        private Features.Wallet.Wallet CreateWallet(string name)
-        {
-            return new Features.Wallet.Wallet()
-            {
-                Name = name,
-                AccountsRoot = new List<AccountRoot>(),
-                BlockLocator = null
-            };
-        }
-
-        private Features.Wallet.Wallet GenerateBlankWallet(string name, string password)
-        {
-            return GenerateBlankWalletWithExtKey(name, password).wallet;
-        }
-
-        private (Features.Wallet.Wallet wallet, ExtKey key) GenerateBlankWalletWithExtKey(string name, string password)
-        {
-            Mnemonic mnemonic = new Mnemonic(Wordlist.English, WordCount.Twelve);
-            ExtKey extendedKey = mnemonic.DeriveExtKey(password);
-
-            Features.Wallet.Wallet walletFile = new Features.Wallet.Wallet
-            {
-                Name = name,
-                EncryptedSeed = extendedKey.PrivateKey.GetEncryptedBitcoinSecret(password, Network.Main).ToWif(),
-                ChainCode = extendedKey.ChainCode,
-                CreationTime = DateTimeOffset.Now,
-                Network = Network.Main,
-                AccountsRoot = new List<AccountRoot> { new AccountRoot { Accounts = new List<HdAccount>(), CoinType = (CoinType)Network.Main.Consensus.CoinType } },
-            };
-
-            return (walletFile, extendedKey);
-        }
-
-        private Block AppendTransactionInNewBlockToChain(ConcurrentChain chain, Transaction transaction)
-        {
-            ChainedBlock last = null;
-            var nonce = RandomUtils.GetUInt32();
-            var block = new Block();
-            block.AddTransaction(transaction);
-            block.UpdateMerkleRoot();
-            block.Header.HashPrevBlock = chain.Tip.HashBlock;
-            block.Header.Nonce = nonce;
-            if (!chain.TrySetTip(block.Header, out last))
-                throw new InvalidOperationException("Previous not existing");
-
-            return block;
-        }
-
-        private Transaction SetupValidTransaction(Features.Wallet.Wallet wallet, string password, HdAddress spendingAddress, PubKey destinationPubKey, HdAddress changeAddress, Money amount, Money fee)
-        {
-            var spendingTransaction = spendingAddress.Transactions.ElementAt(0);
-            Coin coin = new Coin(spendingTransaction.Id, (uint)spendingTransaction.Index, spendingTransaction.Amount, spendingTransaction.ScriptPubKey);
-
-            var privateKey = Key.Parse(wallet.EncryptedSeed, password, wallet.Network);
-
-            var builder = new TransactionBuilder();
-            Transaction tx = builder
-                .AddCoins(new List<Coin> { coin })
-                .AddKeys(new ExtKey(privateKey, wallet.ChainCode).Derive(new KeyPath(spendingAddress.HdPath)).GetWif(wallet.Network))
-                .Send(destinationPubKey.ScriptPubKey, amount)
-                .SetChange(changeAddress.ScriptPubKey)
-                .SendFees(fee)
-                .BuildTransaction(true);
-
-            if (!builder.Verify(tx))
-            {
-                throw new WalletException("Could not build transaction, please make sure you entered the correct data.");
-            }
-
-            return tx;
-        }
-
-        private (ChainedBlock ChainedBlock, Block Block) AppendBlock(ChainedBlock previous, ConcurrentChain chain)
-        {
-            ChainedBlock last = null;
-            var nonce = RandomUtils.GetUInt32();
-            var block = new Block();
-
-            block.AddTransaction(new Transaction());
-            block.UpdateMerkleRoot();
-            block.Header.HashPrevBlock = previous == null ? chain.Tip.HashBlock : previous.HashBlock;
-            block.Header.Nonce = nonce;
-            if (!chain.TrySetTip(block.Header, out last))
-                throw new InvalidOperationException("Previous not existing");
-
-            return (last, block);
-        }
-
-        private void AddAddressesToWallet(WalletManager walletManager, int count)
-        {
-            foreach (var wallet in walletManager.Wallets)
-            {
-                wallet.AccountsRoot.Add(new AccountRoot
-                {
-                    CoinType = CoinType.Bitcoin,
-                    Accounts = new List<HdAccount>
-                    {
-                        new HdAccount
-                        {
-                            ExternalAddresses = GenerateAddresses(count),
-                            InternalAddresses = GenerateAddresses(count)
-                        },
-                        new HdAccount
-                        {
-                            ExternalAddresses = GenerateAddresses(count),
-                            InternalAddresses = GenerateAddresses(count)
-                        } }
-                });
-            }
-        }
-
-        private static HdAddress CreateAddressWithoutTransaction(int index, string addressName)
-        {
-            return new HdAddress()
-            {
-                Index = index,
-                Address = addressName,
-                ScriptPubKey = new Script(),
-                Transactions = new List<TransactionData>()
-            };
-        }
-
-        private static HdAddress CreateAddressWithEmptyTransaction(int index, string addressName)
-        {
-            return new HdAddress()
-            {
-                Index = index,
-                Address = addressName,
-                ScriptPubKey = new Script(),
-                Transactions = new List<TransactionData>() { new TransactionData() }
-            };
-        }
-
-        private List<HdAddress> GenerateAddresses(int count)
-        {
-            List<HdAddress> addresses = new List<HdAddress>();
-            for (int i = 0; i < count; i++)
-            {
-
-                HdAddress address = new HdAddress
-                {
-                    ScriptPubKey = new Key().ScriptPubKey
-                };
-                addresses.Add(address);
-            }
-            return addresses;
-        }
-
-        private static (ExtKey ExtKey, string ExtPubKey) GenerateAccountKeys(Features.Wallet.Wallet wallet, string password, string keyPath)
-        {
-            var accountExtKey = new ExtKey(Key.Parse(wallet.EncryptedSeed, password, wallet.Network), wallet.ChainCode);
-            var accountExtendedPubKey = accountExtKey.Derive(new KeyPath(keyPath)).Neuter().ToString(wallet.Network);
-            return (accountExtKey, accountExtendedPubKey);
-        }
-
-        private static (PubKey PubKey, BitcoinPubKeyAddress Address) GenerateAddressKeys(Features.Wallet.Wallet wallet, string accountExtendedPubKey, string keyPath)
-        {
-            var addressPubKey = ExtPubKey.Parse(accountExtendedPubKey).Derive(new KeyPath(keyPath)).PubKey;
-            var address = addressPubKey.GetAddress(wallet.Network);
-
-            return (addressPubKey, address);
-        }
-
-        private static ConcurrentChain GenerateChainWithHeight(int blockAmount, Network network)
-        {
-            var chain = new ConcurrentChain(network);
-            var nonce = RandomUtils.GetUInt32();
-            var prevBlockHash = chain.Genesis.HashBlock;
-            for (var i = 0; i < blockAmount; i++)
-            {
-                var block = new Block();
-                block.AddTransaction(new Transaction());
-                block.UpdateMerkleRoot();
-                block.Header.HashPrevBlock = prevBlockHash;
-                block.Header.Nonce = nonce;
-                chain.SetTip(block.Header);
-                prevBlockHash = block.GetHash();
-            }
-
-            return chain;
-        }
-
-        private static ConcurrentChain PrepareChainWithBlock()
-        {
-            var chain = new ConcurrentChain(Network.StratisMain);
-            var nonce = RandomUtils.GetUInt32();
-            var block = new Block();
-            block.AddTransaction(new Transaction());
-            block.UpdateMerkleRoot();
-            block.Header.HashPrevBlock = chain.Genesis.HashBlock;
-            block.Header.Nonce = nonce;
-            chain.SetTip(block.Header);
-            return chain;
-        }
-
-        private ICollection<HdAddress> CreateSpentTransactionsOfBlockHeights(Network network, params int[] blockHeights)
-        {
-            var addresses = new List<HdAddress>();
-
-            foreach (int height in blockHeights)
-            {
-                var key = new Key();
-                var address = new HdAddress()
-                {
-                    Address = key.PubKey.GetAddress(network).ToString(),
-                    ScriptPubKey = key.ScriptPubKey,
-                    Transactions = new List<TransactionData>() {
-                        new TransactionData()
-                        {
-                            BlockHeight = height,
-                            Amount = new Money(new Random().Next(500000, 1000000)),
-                            SpendingDetails = new SpendingDetails()
-                        }
-                    }
-                };
-
-                addresses.Add(address);
-            }
-
-            return addresses;
-        }
-
-        private ICollection<HdAddress> CreateUnspentTransactionsOfBlockHeights(Network network, params int[] blockHeights)
-        {
-            var addresses = new List<HdAddress>();
-
-            foreach (int height in blockHeights)
-            {
-                var key = new Key();
-                var address = new HdAddress()
-                {
-                    Address = key.PubKey.GetAddress(network).ToString(),
-                    ScriptPubKey = key.ScriptPubKey,
-                    Transactions = new List<TransactionData>() {
-                        new TransactionData()
-                        {
-                            BlockHeight = height,
-                            Amount = new Money(new Random().Next(500000, 1000000))
-                        }
-                    }
-                };
-
-                addresses.Add(address);
-            }
-
-            return addresses;
-        }
-
         private (Mnemonic mnemonic, Features.Wallet.Wallet wallet) CreateWalletOnDiskAndDeleteWallet(DataFolder dataFolder, string password, string passphrase, string walletName, ConcurrentChain chain)
         {
             var walletManager = new WalletManager(this.LoggerFactory.Object, It.IsAny<ConnectionManager>(), Network.StratisMain, chain, NodeSettings.Default(),
-                                                             dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
+                dataFolder, new Mock<IWalletFeePolicy>().Object, new Mock<IAsyncLoopFactory>().Object, new NodeLifetime());
 
             // create the wallet
             var mnemonic = walletManager.CreateWallet(password, walletName, passphrase);
@@ -3169,46 +2921,6 @@ namespace Stratis.Bitcoin.Tests.Wallet
             File.Delete(dataFolder.WalletPath + $"/{walletName}.wallet.json");
 
             return (mnemonic, wallet);
-        }
-
-        private static TransactionData CreateTransactionDataFromFirstBlock((ConcurrentChain chain, uint256 blockHash, Block block) chainInfo)
-        {
-            var transaction = chainInfo.block.Transactions[0];
-
-            var addressTransaction = new TransactionData()
-            {
-                Amount = transaction.TotalOut,
-                BlockHash = chainInfo.blockHash,
-                BlockHeight = chainInfo.chain.GetBlock(chainInfo.blockHash).Height,
-                CreationTime = DateTimeOffset.FromUnixTimeSeconds(chainInfo.block.Header.Time),
-                Id = transaction.GetHash(),
-                Index = 0,
-                ScriptPubKey = transaction.Outputs[0].ScriptPubKey,
-            };
-            return addressTransaction;
-        }
-
-        public (ConcurrentChain chain, uint256 blockhash, Block block) CreateChainAndCreateFirstBlockWithPaymentToAddress(Network network, HdAddress address)
-        {
-            var chain = new ConcurrentChain(network.GetGenesis().Header);
-
-            Block block = new Block();
-            block.Header.HashPrevBlock = chain.Tip.HashBlock;
-            block.Header.Bits = block.Header.GetWorkRequired(network, chain.Tip);
-            block.Header.UpdateTime(DateTimeOffset.UtcNow, network, chain.Tip);
-
-            var coinbase = new Transaction();
-            coinbase.AddInput(TxIn.CreateCoinbase(chain.Height + 1));
-            coinbase.AddOutput(new TxOut(network.GetReward(chain.Height + 1), address.ScriptPubKey));
-
-            block.AddTransaction(coinbase);
-            block.Header.Nonce = 0;
-            block.UpdateMerkleRoot();
-            block.Header.CacheHashes();
-
-            chain.SetTip(block.Header);
-
-            return (chain, block.GetHash(), block);
         }
     }
 }
